+++ src/router/index.ts
@@ -1,25 +1,25 @@
 import type { App } from 'vue'
 import { createRouter, createWebHashHistory } from 'vue-router';
-import Matrix from '../view/matrix.vue'
-import Flip from '../view/flip.vue'
-import Crop from '../view/crop.vue'
-import Filter from '../view/filter.vue'
-import Color from '../view/color.vue'
-import Mask from '../view/Mask.vue'
+import Matrix from '../view/videoEffect/matrix.vue'
+import Flip from '../view/videoEffect/flip.vue'
+import Crop from '../view/videoEffect/crop.vue'
+import Filter from '../view/videoEffect/filter.vue'
+import Color from '../view/videoEffect/color.vue'
+import Mask from '../view/videoEffect/mask.vue'
 import Hue from '../view/Hue.vue'
-import CirMask from '../view/cir_mask.vue'
+import CirMask from '../view/videoEffect/cir_mask.vue'
 import FiveMask from '../view/five_mask.vue'
-import Transition from '../view/transition.vue'
-import InAnimation from '../view/in_animation.vue'
-import Effect from '../view/effect.vue'
-import Clip from '../view/clip.vue'
-import WebAv from '../view/web_av.vue'
-import WebAvFrame from '../view/web_av_frame.vue'
-import WebAvSpilt from '../view/webav_videoSpilt.vue'
+import Transition from '../view/videoEffect/transition.vue'
+import InAnimation from '../view/videoEffect/in_animation.vue'
+import Effect from '../view/videoEffect/effect.vue'
+import Clip from '../view/webAv/clip.vue'
+import WebAv from '../view/webAv/web_av.vue'
+import WebAvFrame from '../view/webAv/web_av_frame.vue'
+import WebAvSpilt from '../view/webAv/webav_videoSpilt.vue'
 import Index from '../view/index.vue'
-import WebAvVideoAudio from '../view/webav_video_audio.vue'
-import WebAvCanvasAudio from '../view/web_av_canvas_record.vue'
-import WebAvCanvasCamera from '../view/web_av_canvas_camera.vue'
+import WebAvVideoAudio from '../view/webAv/webav_video_audio.vue'
+import WebAvCanvasAudio from '../view/webAv/web_av_canvas_record.vue'
+import WebAvCanvasCamera from '../view/webAv/web_av_canvas_camera.vue'
 import CanvasToVideo from '../view/canvas_to_video.vue'
 import CanvasToMp4 from '../view/canvas_to_mp4.vue'
 import CanvasMp4 from '../view/canvas_mp4.vue'
