+++ src/view/videoEffect/cir_mask.vue
@@ -11,7 +11,7 @@
     
   <script setup lang="ts">
   import { onMounted } from 'vue';
-  import { initShaders } from '../utils/utils'
+  import { initShaders } from '../../utils/utils'
   
   
   const vertexShaderSource = `
