+++ src/view/videoEffect/flip.vue
@@ -9,7 +9,7 @@
 
 <script setup lang="ts">
 import { onMounted } from 'vue';
-import { initShaders } from '../utils/utils'
+import { initShaders } from '../../utils/utils'
 
 
 const vertexShaderSource = `
