+++ src/view/videoEffect/mask.vue
@@ -10,7 +10,7 @@
   
 <script setup lang="ts">
 import { onMounted } from 'vue';
-import { initShaders } from '../utils/utils'
+import { initShaders } from '../../utils/utils'
 
 
 const vertexShaderSource = `
