+++ src/view/videoEffect/transition.vue
@@ -12,7 +12,7 @@
 
 <script lang="ts" setup>
 import { onMounted, ref } from 'vue';
-import { initShaders } from '../utils/utils'
+import { initShaders } from '../../utils/utils'
 
 const vertexShaderSource = `
   attribute vec2 a_Position;
